refactor(phase1a): library.subset を train_util.py から分離
BaseSubset / DreamBoothSubset / FineTuningSubset / ControlNetSubset の 4 クラスを
新規モジュール library/subset.py に移動。train_util.py 側は後方互換のため
re-export shim を残す。

- library/subset.py: 新規（typing 以外の依存なし）
- library/train_util.py: クラス定義 270 行を削除し re-export に置換
- library/config_util.py: subset クラスの import 元を library.subset に切替

動作仕様の変更なし。train_util.py 経由の既存 import は引き続き動作する。
