+++ library/config_util.py
@@ -25,10 +25,12 @@ from voluptuous import (
 from transformers import CLIPTokenizer
 
 from . import train_util
-from .train_util import (
+from .subset import (
     DreamBoothSubset,
     FineTuningSubset,
     ControlNetSubset,
+)
+from .train_util import (
     DreamBoothDataset,
     FineTuningDataset,
     ControlNetDataset,
