+++ library/train_util.py
@@ -409,275 +409,9 @@ class AugHelper:
         return self.color_aug if use_color_aug else None
 
 
-class BaseSubset:
-    def __init__(
-        self,
-        image_dir: Optional[str],
-        alpha_mask: Optional[bool],
-        num_repeats: int,
-        shuffle_caption: bool,
-        caption_separator: str,
-        keep_tokens: int,
-        keep_tokens_separator: str,
-        secondary_separator: Optional[str],
-        enable_wildcard: bool,
-        color_aug: bool,
-        flip_aug: bool,
-        face_crop_aug_range: Optional[Tuple[float, float]],
-        random_crop: bool,
-        caption_dropout_rate: float,
-        caption_dropout_every_n_epochs: int,
-        caption_tag_dropout_rate: float,
-        caption_prefix: Optional[str],
-        caption_suffix: Optional[str],
-        token_warmup_min: int,
-        token_warmup_step: Union[float, int],
-        custom_attributes: Optional[Dict[str, Any]] = None,
-        validation_seed: Optional[int] = None,
-        validation_split: Optional[float] = 0.0,
-        resize_interpolation: Optional[str] = None,
-    ) -> None:
-        self.image_dir = image_dir
-        self.alpha_mask = alpha_mask if alpha_mask is not None else False
-        self.num_repeats = num_repeats
-        self.shuffle_caption = shuffle_caption
-        self.caption_separator = caption_separator
-        self.keep_tokens = keep_tokens
-        self.keep_tokens_separator = keep_tokens_separator
-        self.secondary_separator = secondary_separator
-        self.enable_wildcard = enable_wildcard
-        self.color_aug = color_aug
-        self.flip_aug = flip_aug
-        self.face_crop_aug_range = face_crop_aug_range
-        self.random_crop = random_crop
-        self.caption_dropout_rate = caption_dropout_rate
-        self.caption_dropout_every_n_epochs = caption_dropout_every_n_epochs
-        self.caption_tag_dropout_rate = caption_tag_dropout_rate
-        self.caption_prefix = caption_prefix
-        self.caption_suffix = caption_suffix
-
-        self.token_warmup_min = token_warmup_min  # step=0におけるタグの数
-        self.token_warmup_step = token_warmup_step  # N（N<1ならN*max_train_steps）ステップ目でタグの数が最大になる
-
-        self.custom_attributes = custom_attributes if custom_attributes is not None else {}
-
-        self.img_count = 0
-
-        self.validation_seed = validation_seed
-        self.validation_split = validation_split
-
-        self.resize_interpolation = resize_interpolation
-
-
-class DreamBoothSubset(BaseSubset):
-    def __init__(
-        self,
-        image_dir: str,
-        is_reg: bool,
-        class_tokens: Optional[str],
-        caption_extension: str,
-        cache_info: bool,
-        alpha_mask: bool,
-        num_repeats,
-        shuffle_caption,
-        caption_separator: str,
-        keep_tokens,
-        keep_tokens_separator,
-        secondary_separator,
-        enable_wildcard,
-        color_aug,
-        flip_aug,
-        face_crop_aug_range,
-        random_crop,
-        caption_dropout_rate,
-        caption_dropout_every_n_epochs,
-        caption_tag_dropout_rate,
-        caption_prefix,
-        caption_suffix,
-        token_warmup_min,
-        token_warmup_step,
-        custom_attributes: Optional[Dict[str, Any]] = None,
-        validation_seed: Optional[int] = None,
-        validation_split: Optional[float] = 0.0,
-        resize_interpolation: Optional[str] = None,
-    ) -> None:
-        assert image_dir is not None, "image_dir must be specified / image_dirは指定が必須です"
-
-        super().__init__(
-            image_dir,
-            alpha_mask,
-            num_repeats,
-            shuffle_caption,
-            caption_separator,
-            keep_tokens,
-            keep_tokens_separator,
-            secondary_separator,
-            enable_wildcard,
-            color_aug,
-            flip_aug,
-            face_crop_aug_range,
-            random_crop,
-            caption_dropout_rate,
-            caption_dropout_every_n_epochs,
-            caption_tag_dropout_rate,
-            caption_prefix,
-            caption_suffix,
-            token_warmup_min,
-            token_warmup_step,
-            custom_attributes=custom_attributes,
-            validation_seed=validation_seed,
-            validation_split=validation_split,
-            resize_interpolation=resize_interpolation,
-        )
-
-        self.is_reg = is_reg
-        self.class_tokens = class_tokens
-        self.caption_extension = caption_extension
-        if self.caption_extension and not self.caption_extension.startswith("."):
-            self.caption_extension = "." + self.caption_extension
-        self.cache_info = cache_info
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, DreamBoothSubset):
-            return NotImplemented
-        return self.image_dir == other.image_dir
-
-
-class FineTuningSubset(BaseSubset):
-    def __init__(
-        self,
-        image_dir,
-        metadata_file: str,
-        alpha_mask: bool,
-        num_repeats,
-        shuffle_caption,
-        caption_separator,
-        keep_tokens,
-        keep_tokens_separator,
-        secondary_separator,
-        enable_wildcard,
-        color_aug,
-        flip_aug,
-        face_crop_aug_range,
-        random_crop,
-        caption_dropout_rate,
-        caption_dropout_every_n_epochs,
-        caption_tag_dropout_rate,
-        caption_prefix,
-        caption_suffix,
-        token_warmup_min,
-        token_warmup_step,
-        custom_attributes: Optional[Dict[str, Any]] = None,
-        validation_seed: Optional[int] = None,
-        validation_split: Optional[float] = 0.0,
-        resize_interpolation: Optional[str] = None,
-    ) -> None:
-        assert metadata_file is not None, "metadata_file must be specified / metadata_fileは指定が必須です"
-
-        super().__init__(
-            image_dir,
-            alpha_mask,
-            num_repeats,
-            shuffle_caption,
-            caption_separator,
-            keep_tokens,
-            keep_tokens_separator,
-            secondary_separator,
-            enable_wildcard,
-            color_aug,
-            flip_aug,
-            face_crop_aug_range,
-            random_crop,
-            caption_dropout_rate,
-            caption_dropout_every_n_epochs,
-            caption_tag_dropout_rate,
-            caption_prefix,
-            caption_suffix,
-            token_warmup_min,
-            token_warmup_step,
-            custom_attributes=custom_attributes,
-            validation_seed=validation_seed,
-            validation_split=validation_split,
-            resize_interpolation=resize_interpolation,
-        )
-
-        self.metadata_file = metadata_file
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, FineTuningSubset):
-            return NotImplemented
-        return self.metadata_file == other.metadata_file
-
-
-class ControlNetSubset(BaseSubset):
-    def __init__(
-        self,
-        image_dir: str,
-        conditioning_data_dir: str,
-        caption_extension: str,
-        cache_info: bool,
-        num_repeats,
-        shuffle_caption,
-        caption_separator,
-        keep_tokens,
-        keep_tokens_separator,
-        secondary_separator,
-        enable_wildcard,
-        color_aug,
-        flip_aug,
-        face_crop_aug_range,
-        random_crop,
-        caption_dropout_rate,
-        caption_dropout_every_n_epochs,
-        caption_tag_dropout_rate,
-        caption_prefix,
-        caption_suffix,
-        token_warmup_min,
-        token_warmup_step,
-        custom_attributes: Optional[Dict[str, Any]] = None,
-        validation_seed: Optional[int] = None,
-        validation_split: Optional[float] = 0.0,
-        resize_interpolation: Optional[str] = None,
-    ) -> None:
-        assert image_dir is not None, "image_dir must be specified / image_dirは指定が必須です"
-
-        super().__init__(
-            image_dir,
-            False,  # alpha_mask
-            num_repeats,
-            shuffle_caption,
-            caption_separator,
-            keep_tokens,
-            keep_tokens_separator,
-            secondary_separator,
-            enable_wildcard,
-            color_aug,
-            flip_aug,
-            face_crop_aug_range,
-            random_crop,
-            caption_dropout_rate,
-            caption_dropout_every_n_epochs,
-            caption_tag_dropout_rate,
-            caption_prefix,
-            caption_suffix,
-            token_warmup_min,
-            token_warmup_step,
-            custom_attributes=custom_attributes,
-            validation_seed=validation_seed,
-            validation_split=validation_split,
-            resize_interpolation=resize_interpolation,
-        )
-
-        self.conditioning_data_dir = conditioning_data_dir
-        self.caption_extension = caption_extension
-        if self.caption_extension and not self.caption_extension.startswith("."):
-            self.caption_extension = "." + self.caption_extension
-        self.cache_info = cache_info
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, ControlNetSubset):
-            return NotImplemented
-        return self.image_dir == other.image_dir and self.conditioning_data_dir == other.conditioning_data_dir
+# Subset classes have moved to library.subset; re-exported here for backward compatibility.
+# New code should import from library.subset directly.
+from library.subset import BaseSubset, DreamBoothSubset, FineTuningSubset, ControlNetSubset  # noqa: F401
 
 
 class BaseDataset(torch.utils.data.Dataset):
